requirements.txt
Flask==1.1.1
gunicorn==19.10.0
gevent==1.3.7
flask-expects-json==1.4.0
Flask-Script
redis
flask_caching
pytest-mock
requests
cryptography==3.0
pika==1.1.0